#109 monorepo restructure and rollup downgrade to v3
