+++ Jenkinsfile
@@ -53,7 +53,7 @@ pipeline {
                   dir("${uiKitPath}") {
                       script {
                         echo 'Core packages installing'
-                        sh 'npm ci --legacy-peer-deps'
+                        sh 'npm ci'
                      }
                   }
               }
@@ -65,7 +65,7 @@ pipeline {
                   dir("${rootPath}") {
                       script {
                         echo 'Root packages installing'
-                        sh 'npm ci --legacy-peer-deps'
+                        sh 'npm ci'
                      }
                   }
               }
