+++ core/packages/_tmp-configs/rollup.config.dev.js
@@ -4,7 +4,7 @@ import svg from 'rollup-plugin-image-base64';
 import postcss from 'rollup-plugin-postcss';
 import typescript from 'rollup-plugin-typescript2';
 
-import pkg from './package.json';
+import pkg from '../package.json';
 
 export default {
   input: 'src/index.ts',
