+++ core/packages/_tmp-configs/rollup.config.prod.js--3
@@ -4,7 +4,7 @@ import typescript from 'rollup-plugin-typescript2';
 import terser from '@rollup/plugin-terser';
 import dts from 'rollup-plugin-dts';
 import path from 'path';
-import pkg from './package.json' with { type: 'json' };
+import pkg from './package.json';
 
 export default [
   {
