+++ core/packages/_tmp-configs/rollup.dts.config.js--4
@@ -1,0 +1,8 @@
+import dts from 'rollup-plugin-dts';
+import path from 'path';
+
+export default {
+    input: path.resolve(process.cwd(), 'src/index.tsx'),
+    output: { file: 'dist/index.d.ts', format: 'es' },
+    plugins: [dts()]
+};
