+++ core/packages/package.json
@@ -10,8 +10,8 @@
   "type": "module",
   "scripts": {
     "build-pack": "npm pack",
-    "build": "rollup -c rollup.config.prod.js",
-    "build-dev": "rollup -w -c rollup.config.dev.js",
+    "build": "rollup -c tmp/rollup.config.prod.js",
+    "build-dev": "rollup -w -c tmp/rollup.config.dev.js",
     "prepare-ui-kit": "lerna bootstrap && lerna run build && git reset --hard",
     "clean-node-modules": "rm -rf node_modules"
   },
@@ -21,44 +21,37 @@
   "devDependencies": {
     "@babel/plugin-syntax-jsx": "^7.14.5",
     "@babel/preset-react": "^7.18.6",
-    "@rollup/plugin-commonjs": "29.0.0",
-    "@rollup/plugin-json": "6.1.0",
-    "@rollup/plugin-node-resolve": "16.0.3",
     "@rollup/plugin-terser": "0.4.4",
-    "@rollup/plugin-typescript": "11.1.2",
     "@svgr/rollup": "^6.2.0",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@types/enzyme": "^3.10.10",
-    "@types/jest": "30.0.0",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.1",
     "@types/styled-components": "5.1.26",
-    "@typescript-eslint/parser": "8.46.3",
     "babel-core": "^6.26.3",
     "babel-runtime": "^6.26.0",
     "csstype": "3.0.11",
+    "dayjs": "1.11.19",
     "eslint-config-prettier": "8.6.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "30.2.0",
     "postcss": "8.5.6",
     "postcss-url": "^10.1.3",
     "prettier": "2.8.3",
-    "rimraf": "6.1.0",
-    "rollup": "4.24.0",
-    "rollup-plugin-delete": "2.0.0",
-    "rollup-plugin-dts": "6.1.1",
+    "rollup": "3.29.5",
+    "rollup-plugin-cleaner": "^1.0.0",
+    "rollup-plugin-dts": "6.2.3",
     "rollup-plugin-image-base64": "^1.0.0",
-    "rollup-plugin-postcss": "4.0.2",
+    "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-react-svg": "^3.0.3",
     "rollup-plugin-scss": "^3.0.0",
     "rollup-plugin-svg": "^2.0.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "0.36.0",
     "rollup-pluginutils": "^2.8.2",
-    "tslib": "2.8.1",
     "uuid": "10.0.0"
   },
   "dependencies": {
-    "dayjs": "1.11.19",
     "styled-components": "5.3.3",
     "use-debounce": "10.0.3"
   },
